concert_tour_service/requirements.txt
transformers
torch
sentence-transformers
scikit-learn
numpy
serpapi
python-dotenv
google-search-results
